% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accessibility_2000.R
\name{accessibility_2000}
\alias{accessibility_2000}
\alias{get_accessibility_2000}
\title{Accessibility to Cities in 2000}
\source{
https://forobs.jrc.ec.europa.eu/gam
}
\usage{
get_accessibility_2000()
}
\value{
A function that returns an \code{sf} footprint object.
}
\description{
This resource provides global maps of travel time to cities of 50,000 or more
people in year 2000. Accessibility refers to the ease with which larger cities can be reached from
a certain location. This dataset represents travel time to major cities globally
as of the year 2000, encoded in minutes. The data is essential for historical
analyses, such as understanding the impact of accessibility on land use and
socio-economic outcomes during this period.
}
\references{
European Commission, Joint Research Centre (JRC), Global Accessibility Maps (GAM), 2000.
}
\keyword{resource}
