% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{as_portfolio.list}
\alias{as_portfolio.list}
\title{support for lists of mapme_portfolios e.g. output from group_split}
\usage{
\method{as_portfolio}{list}(ls, atts)
}
\arguments{
\item{ls}{a list of mapme_portfolio objects}
}
\description{
support for lists of mapme_portfolios e.g. output from group_split
}
\keyword{internal}
