% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_backlog.R
\name{available_indicators}
\alias{available_indicators}
\title{Backlog function for available indicators}
\usage{
available_indicators(indicator = NULL)
}
\arguments{
\item{indicator}{Defaults to NULL meaning that a list with all available
indicators will be returned. If a character vector is specified only the
information about the requested indicator will be returned.}
}
\value{
A list object.
}
\description{
This function returns a list of either all available indicators and some
additional metadata or for one or more requested indicators. It can be used
by users of the package to inform themselves about the available data sets
and to learn about potentially additional arguments that should be specified
when requesting the indicator.
}
\examples{
names(available_indicators())
}
\keyword{indicator}
