% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{available_resources}
\alias{available_resources}
\title{Backlog function for available resources}
\usage{
available_resources(resources = NULL)
}
\arguments{
\item{resources}{If \code{NULL} returns a list of all resources (default).
Otherwise only the ones specified.}
}
\value{
A list object.
}
\description{
Returns a list of resource names and parametrization options for the
specified resources. If no resource names are provided, it lists all
available resources, including custom registered ones.
Use it to learn about possible additional arguments that can be specified
when requesting a resource.
}
\examples{
names(available_resources())
}
\keyword{resource}
