% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_avgperc.R
\docType{data}
\name{avgperc}
\alias{avgperc}
\title{Calculate long-term average precipitation}
\format{
A tibble with a column for years, months, the statistic and the coressponding value (in mm)
}
\description{
This functions allows to calculate the monthly long-term average precipitation
based on the CHIRPS data set. The long-term average is defined by the monthly
average precipitation during the climate normal period 1981 - 2010. The average
value of a CHIRPS pixels that intersect with a given polygon are returned.
Users do not need to supply further arguments.
The required resources for this indicator are:
\itemize{
\item \code{chirps}
}
}
\keyword{indicator}
