% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biodiversity_intactness_index.R
\name{biodiversity_intactness_index_resource}
\alias{biodiversity_intactness_index_resource}
\alias{get_biodiversity_intactness_index}
\title{Biodiversity Intactness Index}
\usage{
get_biodiversity_intactness_index(path = NULL)
}
\arguments{
\item{path}{A character vector to the biodiversity intactness index ASCII file.}
}
\value{
A function that returns an \code{sf} footprints object.
}
\description{
The variable is the modeled average abundance of originally-present species,
relative to their abundance in an intact ecosystem. Please refer to
Newbold et al. (2016) for all details, and please cite it when using these
data.
}
\details{
To use this data in mapme workflows, you will have to manually download the
global data set and point towards the file path on your local machine.
Please find the available data under the source link given below.
}
\references{
Tim Newbold; Lawrence Hudson; Andy Arnell; Sara Contu et al.
(2016). Global map of the Biodiversity Intactness Index, from Newbold et al.
(2016) Science [Data set]. Natural History Museum.
\doi{https://doi.org/10.5519/0009936}
}
\keyword{resource}
