% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biome.R
\docType{data}
\name{biome}
\alias{biome}
\title{Calculate biomes statistics (TEOW) based on WWF}
\format{
A tibble with a column for name of the biomes and corresponding area (in ha).
}
\description{
This function allows to efficiently retrieve the name of the biomes
and compute the corresponding area from Terrestrial Ecoregions of the
World (TEOW) - World Wildlife Fund (WWF) for polygons. For each polygon,
the name and area of the biomes (in hectare) is returned.
The required resources for this indicator are:
\itemize{
\item \link{teow}
}
}
\examples{
\dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
}
\dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg",
  package = "mapme.biodiversity"
) \%>\%
  read_sf() \%>\%
  init_portfolio(
    years = 2001,
    outdir = outdir,
    tmpdir = tempdir(),
    add_resources = FALSE,
    verbose = FALSE
  ) \%>\%
  get_resources("teow") \%>\%
  calc_indicators("biome") \%>\%
  tidyr::unnest(biome)

aoi
}
}
\keyword{indicator}
