% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indicators.R
\name{calc_indicators}
\alias{calc_indicators}
\title{Compute specific indicators}
\usage{
calc_indicators(x, ...)
}
\arguments{
\item{x}{An sf portfolio object.}

\item{...}{One or more \verb{calc_*()} functions to calculate indicators for which
the required resources were previously made available via \code{get_resources()}.}
}
\value{
The sf portfolio object \code{x} with additional nested list column per
requested indicator.
}
\description{
With \code{calc_indicators()} specific biodiversity indicators
can be calculated. A requirement is that the resources that
are mandatory inputs for the requested indicators are available
locally. Multiple indicators and their respective additional
arguments can be supplied. You can check available indicators and
their requirement via \code{available_indicators()}, but
the function will also gracefully inform you about any misspecifications.
}
\keyword{function}
