% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_available_years}
\alias{check_available_years}
\title{Helper to check yearly availability}
\usage{
check_available_years(target_years, available_years, indicator)
}
\arguments{
\item{target_years}{Numeric indicating the target year.}

\item{available_years}{Numeric indicating the available years.}

\item{indicator}{A character vector with target resource/indicator name.}
}
\description{
Use this function to check if a specifed vector of years intersects
with the yearly availablity of a resource.
}
\keyword{utils}
