% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{check_engine}
\alias{check_engine}
\title{Check if engine is supported}
\usage{
check_engine(queried_engine)
}
\arguments{
\item{queried_engine}{A character vector of length one indicating the engine
to check for.}
}
\value{
TRUE, invisibly, if the engine is supported. Throws an error otherwise.
}
\description{
Use this functions to check if an extraction engine for zonal vector-raster
operations is supported by the backend.
}
\keyword{utils}
