% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{check_stats}
\alias{check_stats}
\title{Check if statistics are supported}
\usage{
check_stats(queried_stats)
}
\arguments{
\item{queried_stats}{A character vector with statistic names to be checked
if they are supported by the backend}
}
\value{
A character vector of supported statistics Throws an error
if any of the queried statistics is not supported.
}
\description{
Use this function to check if one or multiple statistics are supported
for zonal vector-raster extraction by the backend.
}
\keyword{utils}
