% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_avgperc.R
\name{.calc_avgperc}
\alias{.calc_avgperc}
\title{Calculate long-term precipitation montthly averages}
\usage{
.calc_avgperc(
  shp,
  chirps,
  engine = "zonal",
  rundir = tempdir(),
  verbose = TRUE,
  todisk = FALSE,
  ...
)
}
\arguments{
\item{shp}{A single polygon for which to calculate the tree cover statistic}

\item{chirps}{The CHIRPS resource}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{A directory where intermediate files are written to.}

\item{todisk}{Logical indicating whether or not temporary raster files shall
be written to disk}

\item{...}{additional arguments}
}
\value{
A tibble
}
\description{
Calculate long-term precipitation montthly averages
}
\keyword{internal}
