% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cover.R
\name{.calc_cover}
\alias{.calc_cover}
\title{Calculate tree cover per year based on GFW data sets}
\usage{
.calc_cover(
  shp,
  treecover,
  lossyear,
  minSize = 10,
  minCover = 35,
  rundir = tempdir(),
  verbose = TRUE,
  todisk = FALSE,
  ...
)
}
\arguments{
\item{shp}{A single polygon for which to calculate the tree cover statistic}

\item{treecover}{The treecover 2000 resource from GFW}

\item{lossyear}{The lossyear resource from GFW}

\item{minSize}{The minimum size of a forest patch in ha.}

\item{minCover}{The minimum threshold of stand density for a pixel to be
considered forest in the year 2000.}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{A directory where intermediate files are written to.}

\item{todisk}{Logical indicating whether or not temporary raster files shall
be written to disk}

\item{...}{additional arguments}
}
\value{
A tibble
}
\description{
Considering the 2000 GFW forest cover density layer users can specify
a cover threshold above which a pixel is considered to be covered by forest.
Additionally, users can specify a minimum size of a comprehensive patch of
forest pixels. Patches below this threshold will not be considered as forest
area.
}
