% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dem.R
\name{.calc_dem}
\alias{.calc_dem}
\title{Calculate elevation based on SRTM rasters}
\usage{
.calc_dem(
  shp,
  srtmelevation,
  engine = "zonal",
  stats = "mean",
  rundir = tempdir(),
  verbose = TRUE,
  todisk = FALSE,
  ...
)
}
\description{
Calculate elevation based on SRTM rasters
}
\keyword{internal}
