% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_precipitation.R
\name{.calc_precipitation}
\alias{.calc_precipitation}
\title{Calculate precipitation statistics}
\usage{
.calc_precipitation(
  shp,
  chirps,
  scales_spi = 12,
  engine = "extract",
  rundir = tempdir(),
  verbose = TRUE,
  todisk = FALSE,
  processing_mode = "portfolio",
  ...
)
}
\arguments{
\item{shp}{A single polygon for which to calculate the tree cover statistic}

\item{chirps}{The CHIRPS resource}

\item{scales_spi}{Integers specifying time-scales for SPI}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{A directory where intermediate files are written to.}

\item{todisk}{Logical indicating whether or not temporary raster files shall
be written to disk}

\item{...}{additional arguments}
}
\value{
A tibble
}
\description{
Calculate precipitation statistics
}
\keyword{internal}
