% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_treeloss.R
\name{.calc_treeloss}
\alias{.calc_treeloss}
\title{Calculate treecover are and greenhouse gas emissions}
\usage{
.calc_treeloss(
  shp,
  treecover,
  lossyear,
  greenhouse,
  minSize = 10,
  minCover = 35,
  rundir = tempdir(),
  verbose = TRUE,
  todisk = FALSE
)
}
\arguments{
\item{shp}{A single polygon for which to calculate the tree cover statistic}

\item{treecover}{The treecover 2000 resource from GFW}

\item{lossyear}{The lossyear resource from GFW}

\item{greenhouse}{The greenhouse emission layer from GFW}

\item{minSize}{The minimum size of a forest patch in ha.}

\item{minCover}{The minimum threshold of stand density for a pixel to be considered forest in the year 2000.}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{A directory where intermediate files are written to.}

\item{todisk}{Logical indicating whether or not temporary raster files shall be written to disk}
}
\value{
A tibble
}
\description{
This is a joint indicator calculating both, treecover and greenhouse
gas emissions per treecover loss. It actually combines the indicators for
treecover and emissions into a single function since the computations for both
indicators are very similar. If a user is interested only in one of these
two indicators consider applying the respective functions.
}
