% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_available_years}
\alias{.check_available_years}
\title{Helper to check year availability}
\usage{
.check_available_years(target_years, available_years, indicator)
}
\arguments{
\item{target_years}{Numeric/s indicating the target year/s}

\item{available_years}{Numeric/s indicating the available year/s}

\item{indicator}{A character vector with target indicator}
}
\description{
Helper to check year availability
}
\keyword{internal}
