% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_existing_resources}
\alias{.check_existing_resources}
\title{Helper to check for existing resources}
\usage{
.check_existing_resources(ex_resources, req_resources, needed = FALSE)
}
\arguments{
\item{ex_resources}{A character vector with existing resources}

\item{req_resources}{A character vector with requested resources}

\item{needed}{default value being FALSE}
}
\description{
Helper to check for existing resources
}
\keyword{internal}
