% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_requested_indicator}
\alias{.check_requested_indicator}
\title{Helper to check for supported indicators}
\usage{
.check_requested_indicator(indicators)
}
\arguments{
\item{indicators}{A character vector with requested indicators}
}
\description{
Helper to check for supported indicators
}
\keyword{internal}
