% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_requested_resources}
\alias{.check_requested_resources}
\title{Helper to check for supported resources}
\usage{
.check_requested_resources(resources)
}
\arguments{
\item{resources}{A character vector with requested resources}
}
\description{
Helper to check for supported resources
}
\keyword{internal}
