% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_resource_arguments}
\alias{.check_resource_arguments}
\title{Helper to check for resource arguments}
\usage{
.check_resource_arguments(resource, args)
}
\arguments{
\item{resource}{A character vector with requested resource}

\item{args}{The arguments of requested resource}
}
\description{
Helper to check for resource arguments
}
\keyword{internal}
