% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dem.R
\name{.comp_dem_exact_extractr}
\alias{.comp_dem_exact_extractr}
\title{function to compute single raster with exactextractr}
\usage{
.comp_dem_exact_extractr(elevation = NULL, shp = NULL, stats = "mean", ...)
}
\description{
function to compute single raster with exactextractr
}
\keyword{internal}
