% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dem.R
\name{.comp_dem_extract}
\alias{.comp_dem_extract}
\title{function to compute single raster with terra extract}
\usage{
.comp_dem_extract(elevation = NULL, shp = NULL, stats = "mean", ...)
}
\description{
function to compute single raster with terra extract
}
\keyword{internal}
