% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dem.R
\name{.comp_dem_zonal}
\alias{.comp_dem_zonal}
\title{function to compute single raster with terra zonal}
\usage{
.comp_dem_zonal(
  elevation = NULL,
  shp = NULL,
  stats = "mean",
  todisk = FALSE,
  rundir = tempdir(),
  ...
)
}
\description{
function to compute single raster with terra zonal
}
\keyword{internal}
