% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tri.R
\name{.comp_tri_exact_extractr}
\alias{.comp_tri_exact_extractr}
\title{Helper function to compute statistics using routines from exactextractr}
\usage{
.comp_tri_exact_extractr(
  elevation = NULL,
  shp = NULL,
  stats = "mean",
  todisk = todisk,
  rundir = rundir,
  ...
)
}
\arguments{
\item{elevation}{elevation raster from which to compute statistics}
}
\value{
A data-frame
}
\description{
Helper function to compute statistics using routines from exactextractr
}
\keyword{internal}
