% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tri.R
\name{.comp_tri_extract}
\alias{.comp_tri_extract}
\title{function to compute single raster with terra extract}
\usage{
.comp_tri_extract(
  elevation = NULL,
  shp = NULL,
  stats = "mean",
  todisk = todisk,
  rundir = rundir,
  ...
)
}
\description{
function to compute single raster with terra extract
}
\keyword{internal}
