% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tri.R
\name{.comp_tri_zonal}
\alias{.comp_tri_zonal}
\title{Helper function to compute statistics using routines from terra zonal}
\usage{
.comp_tri_zonal(
  elevation = NULL,
  shp = NULL,
  stats = "mean",
  todisk = FALSE,
  rundir = tempdir,
  ...
)
}
\arguments{
\item{elevation}{elevation raster from which to compute statistics}
}
\value{
A data-frame
}
\description{
Helper function to compute statistics using routines from terra zonal
}
\keyword{internal}
