% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.download_or_skip}
\alias{.download_or_skip}
\title{Helper to check valid urls}
\usage{
.download_or_skip(
  urls,
  filenames,
  verbose,
  stubbornnes = 6,
  check_existence = TRUE
)
}
\arguments{
\item{urls}{address to the datasource}

\item{filenames}{name of the files in the url}

\item{verbose}{Logical controlling verbosity.}

\item{stubbornnes}{default value being 6}

\item{check_existence}{default to TRUE}
}
\description{
Helper to check valid urls
}
\keyword{internal}
