% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accessibility.R
\name{.get_accessibility}
\alias{.get_accessibility}
\title{Downloads Accessibility to Cities layer}
\usage{
.get_accessibility(
  x,
  range_accessibility = "20k_50k",
  rundir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{range_accessibility}{The city within the defined range of population
to download, defaults to \code{"20k_50k"}.}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
Downloads Accessibility to Cities layer
}
\keyword{internal}
