% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accessibility.R
\name{.get_accessibility_url}
\alias{.get_accessibility_url}
\title{Helper for accessibility urls generation}
\usage{
.get_accessibility_url(range)
}
\arguments{
\item{range}{A valid range that is translated to an url}
}
\value{
A character string
}
\description{
Helper for accessibility urls generation
}
\keyword{internal}
