% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climaticVariables.R
\name{.get_climate_url}
\alias{.get_climate_url}
\title{Helper function to construct climate variable layers urls}
\usage{
.get_climate_url(layer, year)
}
\arguments{
\item{layer}{A character indicating the target variable name}

\item{year}{A numeric indicating the target year}
}
\value{
A character vector.
}
\description{
Helper function to construct climate variable layers urls
}
\keyword{internal}
