% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climaticVariables.R
\name{.get_climatic_variables}
\alias{.get_climatic_variables}
\title{Helper function to download climate variable layers}
\usage{
.get_climatic_variables(x, layer, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{layer}{A character indicating the target variable name}
}
\value{
A character vector.
}
\description{
Helper function to download climate variable layers
}
\keyword{internal}
