% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_droughtInd.R
\name{.get_drought_ind}
\alias{.get_drought_ind}
\title{Downloads NASA GRACE-based Drought Indicator Layer}
\usage{
.get_drought_ind(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
Downloads NASA GRACE-based Drought Indicator Layer
}
\keyword{internal}
