% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_droughtInd.R
\name{.get_drought_url}
\alias{.get_drought_url}
\title{Helper function to construct GRACE urls}
\usage{
.get_drought_url(target_year)
}
\arguments{
\item{target_year}{}
}
\value{
A character vector
}
\description{
Helper function to construct GRACE urls
}
\keyword{internal}
