% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecoregions.R
\name{.get_ecoregions}
\alias{.get_ecoregions}
\title{Downloads Terrestrial Ecoregions of the World (TEOW) Polygon}
\usage{
.get_ecoregions(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
Downloads Terrestrial Ecoregions of the World (TEOW) Polygon
}
\keyword{internal}
