% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ESALandCover.R
\name{.get_esa_land_cover}
\alias{.get_esa_land_cover}
\title{Downloads Copernicus Global Land Cover layer}
\usage{
.get_esa_land_cover(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
Downloads Copernicus Global Land Cover layer
}
\keyword{internal}
