% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ESALandCover.R
\name{.get_esa_url}
\alias{.get_esa_url}
\title{Helper function to create ESA land cover urls}
\usage{
.get_esa_url(tile, year)
}
\arguments{
\item{tile}{An sf object representing the spatial extent of the a tile}

\item{year}{A single numeric value indicating the target year}
}
\value{
A character vector
}
\description{
Helper function to create ESA land cover urls
}
\keyword{internal}
