% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_greenhouse.R
\name{.get_greenhouse}
\alias{.get_greenhouse}
\title{Get greenhouse layer}
\usage{
.get_greenhouse(x, verbose = TRUE, rundir = tempdir())
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{verbose}{Logical controlling verbosity.}

\item{rundir}{A directory where intermediate files are written to.}
}
\description{
Get greenhouse layer
}
\keyword{internal}
