% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lossyear.R
\name{.get_lossyear}
\alias{.get_lossyear}
\title{Get lossyear layer}
\usage{
.get_lossyear(
  x,
  vers_lossyear = "GFC-2020-v1.8",
  rundir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{vers_lossyear}{The version to download, defaults to
\code{"GFC-2020-v1.8"}.}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{A directory where intermediate files are written to.}
}
\description{
Get lossyear layer
}
\keyword{internal}
