% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mangrove.R
\name{.get_mangrove}
\alias{.get_mangrove}
\title{Downloads Global Mangrove Extent Polygon}
\usage{
.get_mangrove(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
Downloads Global Mangrove Extent Polygon
}
\keyword{internal}
