% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mangrove.R
\name{.get_mangrove_url}
\alias{.get_mangrove_url}
\title{A helper function to construct correct mangrove layer urls}
\usage{
.get_mangrove_url(target_year)
}
\arguments{
\item{target_year}{A numeric indicating the target year}
}
\value{
A character vector
}
\description{
A helper function to construct correct mangrove layer urls
}
\keyword{internal}
