% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasagrace.R
\name{.get_nasagrace_url}
\alias{.get_nasagrace_url}
\title{Helper function to construct GRACE urls}
\usage{
.get_nasagrace_url(target_year)
}
\arguments{
\item{target_year}{}
}
\value{
A character vector
}
\description{
Helper function to construct GRACE urls
}
\keyword{internal}
