% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_popCount.R
\name{.get_popCount}
\alias{.get_popCount}
\title{Downloads Population Count layer}
\usage{
.get_popCount(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
Downloads Population Count layer
}
\keyword{internal}
