% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indicator.R
\name{.get_single_indicator}
\alias{.get_single_indicator}
\title{Calculation of an indicator}
\usage{
.get_single_indicator(x, indicator, ...)
}
\arguments{
\item{x}{A sf object returned by init_portfolio().}

\item{indicator}{A variable length charcter vector with the indicators to
calculate.}

\item{...}{Additional arguments required by the requested indicators.}
}
\description{
This functions let's users calculate on or more biodiversity indicators for
a portfolio.
}
\keyword{internal}
