% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SRTMdem.R
\name{.get_srtm_dem}
\alias{.get_srtm_dem}
\title{Downloads SRTM 30m Digital Elevation Model (DEM) Layer}
\usage{
.get_srtm_dem(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
Downloads SRTM 30m Digital Elevation Model (DEM) Layer
}
\keyword{internal}
