% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SRTMdem.R
\name{.get_srtm_url}
\alias{.get_srtm_url}
\title{Helper function to construct SRTM urls}
\usage{
.get_srtm_url(tile)
}
\arguments{
\item{tile}{An sf object representing the tile to download}
}
\value{
A character vector
}
\description{
Helper function to construct SRTM urls
}
\keyword{internal}
