% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treecover.R
\name{.get_treecover}
\alias{.get_treecover}
\title{Get treecover layer}
\usage{
.get_treecover(
  x,
  vers_treecover = "GFC-2020-v1.8",
  rundir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{vers_treecover}{The version to download, defaults to
\code{"GFC-2020-v1.8"}.}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
Get treecover layer
}
\keyword{internal}
