% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accessibility.R
\name{.getaccessibilityURL}
\alias{.getaccessibilityURL}
\title{Helper for accessibility urls generation}
\usage{
.getaccessibilityURL(range)
}
\arguments{
\item{range}{A valid range that is translated to an url}
}
\value{
A character string
}
\description{
Helper for accessibility urls generation
}
\keyword{internal}
