% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mangrove.R
\name{.getMangroveURL}
\alias{.getMangroveURL}
\title{A helper function to construct correct mangrove layer urls}
\usage{
.getMangroveURL(target_year)
}
\arguments{
\item{target_year}{A numeric indicating the target year}
}
\description{
A helper function to construct correct mangrove layer urls
}
\keyword{internal}
