% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_popCount.R
\name{.getPopCountURL}
\alias{.getPopCountURL}
\title{Helper function to construct population layer urls}
\usage{
.getPopCountURL(target_year)
}
\arguments{
\item{target_year}{A numeric indicating the target year}
}
\value{
A character vector.
}
\description{
Helper function to construct population layer urls
}
\keyword{internal}
