% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SRTMdem.R
\name{.getSrtmURL}
\alias{.getSrtmURL}
\title{Helper function to construct SRTM urls}
\usage{
.getSrtmURL(tile)
}
\arguments{
\item{tile}{An sf object representing the tile to download}
}
\value{
A charchter vector
}
\description{
Helper function to construct SRTM urls
}
\keyword{internal}
