% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_indicator.R
\name{.prep_and_compute}
\alias{.prep_and_compute}
\title{Internal indicator routine}
\usage{
.prep_and_compute(shp, params, i)
}
\description{
Helper to abstract preparation and computation
of indicators per polygon
}
\keyword{internal}
