% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.unzip_and_remove}
\alias{.unzip_and_remove}
\title{Helper to unzip and remove zip files}
\usage{
.unzip_and_remove(zip, rundir, remove = TRUE)
}
\arguments{
\item{zip}{zip file to unzip}

\item{rundir}{A directory where intermediate files are written to.}

\item{remove}{if TRUE, removes the zip else keeps it}
}
\description{
Helper to unzip and remove zip files
}
\keyword{internal}
