% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_drought_indicator.R
\docType{data}
\name{drought_indicator}
\alias{drought_indicator}
\title{Calculate drought indicator statistics}
\format{
A tibble with a column for each specified stats and a column with the respective date.
}
\description{
This function allows to efficiently calculate drought indicator statistics for
polygons. For each polygon, the desired statistic/s (mean, median or sd) is/are
returned. The required resources for this indicator are:
\itemize{
\item \code{droughtindicators}
}
}
\details{
The following arguments can be set:
\describe{
\item{stats_drought}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character "mean", "median" or "sd".}
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}
}
}
\keyword{indicator}
