% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_elevation.R
\name{elevation}
\alias{elevation}
\alias{calc_elevation}
\title{Calculate elevation statistics}
\usage{
calc_elevation(stats = "mean", engine = NULL)
}
\arguments{
\item{stats}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character "mean", "median" or "sd".}

\item{engine}{Deprecated. Will be removed in a future release.}
}
\value{
A function that returns an indicator tibble with specified elevation
statistics as variable and corresponding values (in meters) as value.
}
\description{
This function allows to calculate elevation statistics for
polygons. For each polygon, the desired statistic(s)
are returned.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{nasa_srtm}
}
}
\examples{
\dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
}
\dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

mapme_options(
  outdir = outdir,
  verbose = FALSE
)

aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg",
  package = "mapme.biodiversity"
) \%>\%
  read_sf() \%>\%
  get_resources(get_nasa_srtm()) \%>\%
  calc_indicators(
    calc_elevation(stats = c("mean", "median", "stdev", "var"))
  ) \%>\%
  portfolio_long()

aoi
}
}
\keyword{indicator}
