% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ESALandCover.R
\name{ESA_Global_Land_Cover}
\alias{ESA_Global_Land_Cover}
\alias{.get_esa_land_cover}
\title{Downloads Copernicus Global Land Cover layer}
\usage{
.get_esa_land_cover(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
The resource represents the actual surface cover of ground available
annually for the period 2015 to 2019. The cell values range from 0 to
200, representing total of 23 discrete classifications from ESA.
}
\keyword{internal}
