% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_firecci51.R
\name{firecci51}
\alias{firecci51}
\alias{get_firecci51}
\title{FireCCI51 Burned Area Product}
\source{
\url{https://data.ceda.ac.uk/neodc/esacci/fire/data/burned_area/MODIS/pixel/v5.1/}
}
\usage{
get_firecci51(years = 2001:2020)
}
\arguments{
\item{years}{Numeric vector of years to make the FireCCI51 product available for. Must be greater than the year 2001.}
}
\value{
A function that returns an \code{sf} footprint object.
}
\description{
The FireCCI51 product provides global burned area maps at 250m resolution from 2001 onwards.
The data is produced by the European Space Agency (ESA) and is available on the CEDA platform.
}
\references{
Alonso-Canas, I. and Chuvieco, E. (2015). Global Burned Area Mapping from MODIS (2001-2016). \doi{10.1016/j.rse.2015.08.009}
}
\keyword{resource}
