% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treecover.R
\name{Forest_Treecover_2000}
\alias{Forest_Treecover_2000}
\alias{.get_treecover}
\title{Downloads Tree Cover 2000 layer}
\usage{
.get_treecover(
  x,
  vers_treecover = "GFC-2020-v1.8",
  rundir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{vers_treecover}{The version to download, defaults to
\code{"GFC-2020-v1.8"}.}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
This resource represents the tree cover in the year 2000, defined as canopy
closure for all vegetation taller than 5m in height. It is encoded as a
percentage per output grid cell in the range 0 to 100.
}
\keyword{internal}
