% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fritz_et_al.R
\name{fritz_et_al}
\alias{fritz_et_al}
\alias{get_fritz_et_al}
\title{Drivers of deforestation for tropical forests}
\source{
\url{https://zenodo.org/record/7997885}
}
\usage{
get_fritz_et_al(resolution = 100)
}
\arguments{
\item{resolution}{An integer indicating the resolution
to download. Defaults to 100.}
}
\value{
A function that returns an \code{sf} footprint object.
}
\description{
This resource is produced by a neirest-neighbour matching of a crowd-sourced
campaign to map dominant driver of forest loss based on visual interpretation
of VHR images matched with Global Forest Loss data by Hansen (2013) version
1.7 The forest loss layer was re sampled to a resolution of 100 and 1.000
meters. Dominant drivers were determined for the period 2008 to 2009.
}
\details{
It indicates 9 different classes:
\itemize{
\item commercial agriculture
\item commercial oil palm plantations
\item managed forests
\item mining
\item natural disturbances
\item pasture
\item roads
\item wildfire
\item other subsistence agriculture
\item shifting cultivation
}
}
\references{
Steffen, F., Carlos, J.C.L., See. L., Schepaschenko D.,
Hofhansl F., Jung M., Dürauer M., Georgieva I., Danylo O., Lesiv M.,
McCallum I. (2022) A Continental Assessment of the Drivers of Tropical
Deforestation With a Focus on Protected Areas. F.Cos.Sc.(3)
doi:10.3389/fcosc.2022.830248
}
\keyword{resource}
