% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resources.R
\name{get_resources}
\alias{get_resources}
\title{Download biodiversity resources}
\usage{
get_resources(x, ...)
}
\arguments{
\item{x}{A biodiversity portfolio object constructed via
\code{init_portfolio()}}

\item{...}{One or more \verb{get_*()} functions to make a resource available.}
}
\value{
Called for its side effect of making resources available in the
package environment. Returns x, invisibly.
}
\description{
With \code{get_resources()} data sets required for the
calculation of indicators can be made available. The function supports the
specification of several resource functions. To determine the output path,
temporary directory and verbosity, the output of \code{mapme_options()} is used.
}
\keyword{function}
