% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghsl.R
\docType{data}
\name{ghsl}
\alias{ghsl}
\alias{get_ghsl}
\title{Population Count layer (GHSL)}
\source{
\url{https://ghsl.jrc.ec.europa.eu/ghs_pop2023.php}
}
\usage{
get_ghsl(years = 2020)
}
\arguments{
\item{years}{A numeric vector indicating the years for which to make the
resource available. Supported years are: 1975, 1980, 1985, 1990, 1995,
2000, 2005, 2010, 2015, 2020, 2025, 2030.}
}
\value{
A function that returns an \code{sf} footprint object.
}
\description{
The Global Human Settlement Layer (GHSL) provides a set of global spatial
information, evidence-based analytics, and knowledge about the human
presence on the planet. This resource represents the population count,
30 arcsec (~1 km) spatial resolution layers available to download for
the years 1975 to 2030 in 5-year intervals.
The dataset used is the GHS-POP R2023A version in WGS84 (EPSG:4326).
}
\keyword{resource}
