% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ghsl_pop.R
\name{ghsl_pop}
\alias{ghsl_pop}
\alias{get_ghsl_pop}
\title{GHS-POP Population Grid for 1975-2030}
\source{
\url{https://human-settlement.emergency.copernicus.eu/ghs_pop2023.php}
}
\usage{
get_ghsl_pop(years = 2020)
}
\arguments{
\item{years}{A numeric vector indicating the years for which to make the
resource available. Must be multiples of 5 between 1975 and 2030.}
}
\value{
A function that returns an \code{sf} footprint object.
}
\description{
This resource is published by the European Commission Joint Research Centre
(JRC) as part of the Global Human Settlement Layer (GHSL) project. It
represents the distribution of resident population, expressed as the number
of people per cell, at 30 arc-second (~1 km) spatial resolution in WGS84
(EPSG:4326). Population estimates are derived from CIESIN GPWv4.11 and
disaggregated using the built-up surface and volume information from the
GHSL. Data are available at 5-year intervals from 1975 to 2030.
}
\details{
It may be required to increase the timeout option to successfully download
these layers from their source location via e.g.
\code{options(timeout = 600)}.
}
\references{
Pesaresi, M. et al. (2024). Advances on the Global Human
Settlement Layer by joint assessment of Earth Observation and population
survey data. International Journal of Digital Earth, 17(1).
\doi{doi:10.1080/17538947.2024.2390454}
}
\keyword{resource}
