% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ghsl_population.R
\name{ghsl_population}
\alias{ghsl_population}
\alias{calc_ghsl_population}
\title{Calculate GHSL population count statistics}
\usage{
calc_ghsl_population(engine = "extract", stats = "sum")
}
\arguments{
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}

\item{stats}{Function to be applied to compute statistics for polygons
either one or multiple inputs as character "min", "max", "sum", "mean",
"median" "sd" or "var".}
}
\value{
A function that returns an indicator tibble with the specified
population statistics as variable and the corresponding values as value.
}
\description{
The Global Human Settlement Layer (GHSL) provides population grids derived
from CIESIN GPWv4.11, disaggregated using built-up surface and volume
information. This function allows to efficiently calculate population count
statistics (e.g. total number of population) for polygons. For each polygon,
the desired statistic/s (min, max, sum, mean, median, sd or var) is/are
returned.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{ghsl_pop}
}
}
\examples{
\dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
}
\dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

mapme_options(
  outdir = outdir,
  verbose = FALSE
)

aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg",
  package = "mapme.biodiversity"
) \%>\%
  read_sf() \%>\%
  get_resources(get_ghsl_pop(years = c(2000, 2010, 2020))) \%>\%
  calc_indicators(
    calc_ghsl_population(engine = "extract", stats = c("sum", "median"))
  ) \%>\%
  portfolio_long()

aoi
}
}
\keyword{indicator}
