% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mangrove.R
\name{Global_Mangrove_Extent}
\alias{Global_Mangrove_Extent}
\alias{.get_mangrove}
\title{Downloads Global Mangrove Extent Polygon}
\usage{
.get_mangrove(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
The polygons represent the mangrove, which is tropical coastal vegetation
and considered the most significant part of the marine ecosystem. This
resource is available for the period 1996- 2016 from Global Mangrove Watch
(GMW), providing geospatial information about global mangrove extent.
}
\keyword{internal}
