% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsw.R
\docType{data}
\name{global_surface_water_change}
\alias{global_surface_water_change}
\title{Global Surface Water Change}
\format{
Global raster layers
}
\source{
\url{https://global-surface-water.appspot.com/}
}
\description{
The change in water occurrence intensity between the two periods is derived
from homologous pairs of months (i.e. same months containing valid
observations in both periods). The difference in the occurrence of surface
water was calculated for each homologous pair of months. The average of all
of these differences constitutes the Surface Water Occurrence change
intensity. The raster files have integer cell values between \code{[0, 200]}
where 0 represents surface water loss and 200 represents surface water gain.
}
\keyword{resource}
