% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsw.R
\docType{data}
\name{global_surface_water_recurrence}
\alias{global_surface_water_recurrence}
\title{Global Surface Water Recurrence}
\format{
Global raster layers
}
\source{
\url{https://global-surface-water.appspot.com/}
}
\description{
Water Recurrence is a measurement of the degree of variability in the
presence of water from year to year. It describes the frequency with which
water returned to a particular location from one year to another, and is
expressed as a percentage. The raster files have integer cell values between
\code{[0, 100]}, where 100 represents that water reoccurs predictably every
year, whereas lower values indicate that water only occurs episodically.
}
\keyword{resource}
