% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsw.R
\docType{data}
\name{global_surface_water_seasonality}
\alias{global_surface_water_seasonality}
\title{Global Surface Water Seasonality}
\format{
Global raster layers
}
\source{
\url{https://global-surface-water.appspot.com/}
}
\description{
GSW seasonality describes the intra-annual distribution of surface water for
each pixel. The raster files have integer cell values between \code{[0, 12]},
indicating how many months per year the pixel was classified as water.
}
\keyword{resource}
