% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsw.R
\docType{data}
\name{global_surface_water_transitions}
\alias{global_surface_water_transitions}
\title{Global Surface Water Transitions}
\format{
Global raster layers
}
\source{
\url{https://global-surface-water.appspot.com/}
}
\description{
GSW transition data contains information about the type of surface water
change for each pixel. The raster files have integer cell values between
\code{[0, 10]} that code for different transition classes:
}
\details{
\tabular{ll}{
   Value \tab Transition Class \cr
   1 \tab Permanent \cr
   2 \tab New Permanent \cr
   3 \tab Lost Permanent \cr
   4 \tab Seasonal \cr
   5 \tab New Seasonal \cr
   6 \tab Lost Seasonal \cr
   7 \tab Seasonal to Permanent \cr
   8 \tab Permanent to Seasonal \cr
   9 \tab Ephemeral Permanent \cr
   10 \tab Ephemeral Seasonal \cr
}
}
\keyword{resource}
