% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gmw.R
\docType{data}
\name{gmw}
\alias{gmw}
\title{Calculate mangrove extent based on Global Mangrove Watch (GMW)}
\format{
A tibble with a column for area of mangrove (in ha) and corresponding year.
}
\description{
This function allows to efficiently calculate area of mangrove from
Global Mangrove Watch - World Conservation Monitoring Centre (WCMC)
for polygons. For each polygon, the area of the mangrove (in hectare)
for desired year is returned.
The required resources for this indicator are:
\itemize{
\item \link{mangrove}
}
}
\examples{
library(sf)
library(mapme.biodiversity)
(aoi <- system.file("extdata", "shell_beach_protected_area_41057_B.gpkg",
  package = "mapme.biodiversity"
) \%>\%
  read_sf() \%>\%
  init_portfolio(
    years = c(1996, 2016),
    outdir = system.file("res", package = "mapme.biodiversity"),
    tmpdir = system.file("tmp", package = "mapme.biodiversity"),
    add_resources = FALSE,
    cores = 1,
    verbose = FALSE
  ) \%>\%
  get_resources("mangrove") \%>\%
  calc_indicators("gmw") \%>\%
  tidyr::unnest(gmw))
}
\keyword{indicator}
