% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gsw_change.R
\docType{data}
\name{gsw_change}
\alias{gsw_change}
\title{Calculate Global Surface Water (GSW) Change}
\format{
A tibble with a column for the aggregated GSW change indicator.
}
\description{
The change in water occurrence intensity between the two periods is derived
from homologous pairs of months (i.e. same months containing valid
observations in both periods). The difference in the occurrence of surface
water was calculated for each homologous pair of months. The average of all
of these differences constitutes the Surface Water Occurrence change
intensity. The raster files have integer cell values between \code{[0, 200]}
where 0 represents surface water loss and 200 represents surface water gain.
}
\details{
The pixel values are aggregated using method provided via the
\code{stats_gsw} parameter.

The required resources for this indicator are:
\itemize{
\item \link{global_surface_water_change}
The following arguments can be set:
\describe{
\item{stats_gsw}{The aggregation function applied to the input raster
values. Defaults to \code{mean}.}
}
}
}
\examples{
\dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
}
\dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

aoi <- system.file("extdata", "shell_beach_protected_area_41057_B.gpkg",
  package = "mapme.biodiversity"
) \%>\%
  read_sf() \%>\%
  init_portfolio(
    years = 2001,
    outdir = outdir,
    tmpdir = tempdir(),
    verbose = FALSE
  ) \%>\%
  get_resources("global_surface_water_change") \%>\%
  calc_indicators("gsw_change")

aoi
}
}
\keyword{indicator}
