% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gsw_time_series.R
\name{gsw_time_series}
\alias{gsw_time_series}
\alias{calc_gsw_time_series}
\title{Calculate Global Surface Water Time Series}
\format{
A function returning a tibble with time series of global surface
water data classes.
}
\usage{
calc_gsw_time_series(years = 1984:2021)
}
\arguments{
\item{years}{Numeric vector of years to process between 1984 and 2021.
Default: \code{1984:2021}.}
}
\description{
This function calculates the total area of the global surface water time
series data, separated by the following classes:
}
\details{
\itemize{
\item No Observation: It was not possible to determine whether a pixel was water
(this may be the case for frozen areas or during the polar night in extreme
latitudes).
\item Permanent Water: Water was detected in twelve months per year or in a
combination of permanent and no observation.
\item Seasonal Water: Water and no water was detected.
\item No Water: No Water was detected.
}

The required resources for this indicator are:
\itemize{
\item \link{gsw_time_series}
}
}
\keyword{indicator}
