% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{init_portfolio}
\alias{init_portfolio}
\title{Conversion method for sf objects}
\usage{
init_portfolio(x, years, outdir, tmpdir = tempdir())
}
\arguments{
\item{x}{The sf object to be transformed to a portfolio}

\item{years}{Numeric vector for the analysis years}

\item{outdir}{Character indicating the directory where resources are written}

\item{tmpdir}{Character vector to a directory used for intermediate files.
Defaults to the output of tempdir(). Should not be identical with outdir.}
}
\description{
Checks to be performed:
\itemize{
\item has at least one asset
\item all assets are either Polygon or MultiPolygon (TODO)
\item is Lat/Lon, otherwise transform to EPSG:4326
}
}
