% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iucn.R
\name{iucn}
\alias{iucn}
\alias{get_iucn}
\title{IUCN Red List of Threatened Species}
\source{
\url{https://www.iucnredlist.org/resources/other-spatial-downloads}
}
\usage{
get_iucn(paths = NULL)
}
\arguments{
\item{paths}{A character vector to the respective species range files in GTiff
format. Note, that theses files have to be downloaded manually.}
}
\value{
A function that returns an \code{sf} footprint object.
}
\description{
This resource is part of the spatial data set Red List of Threatened Species
released by IUCN. It is free to use under a non-commercial licence. For
commercial uses, a request has to be sent to Integrated Biodiversity
Assessment Tool (IBAT).
}
\details{
To use this data in mapme workflows, you will have to manually download the
global data set and point towards the file path on your local machine.
Please find the available data under the source link given below.
}
\references{
IUCN (2024). The IUCN Red List of Threatened Species.
\url{https://www.iucnredlist.org}
}
\keyword{resource}
