% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_key_biodiversity_areas.R,
%   R/get_key_biodiversity_areas.R
\docType{data}
\name{key_biodiversity_areas}
\alias{key_biodiversity_areas}
\alias{calc_key_biodiversity_area}
\alias{get_key_biodiversity_areas}
\title{Calculate Key Biodiversity Areas}
\format{
A function returning an indicator tibble with \code{key_biodiversity_area}
as variable and the total overlap area (in ha) as value.
}
\source{
\url{https://www.keybiodiversityareas.org/kba-data}
}
\usage{
calc_key_biodiversity_area()

get_key_biodiversity_areas(path = NULL)
}
\arguments{
\item{path}{A character vector to the key biodiversity areas GPKG file.
Note, that the file has to be downloaded manually.}
}
\value{
A function that returns an \code{sf} footprints object.
}
\description{
This function calculates the total area of key biodiversity areas for a given
input polygon.

This resource contains outlines of key biodiversity areas, which are areas
representing sites with specific importance for nature conservation.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{key_biodiversity_areas}
}

To use this data in mapme workflows, you will have to manually download the
global data set and point towards its file path on your local machine.
Please find the available data under the source link given below.
}
\examples{
\dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
}
\dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

mapme_options(
  outdir = outdir,
  verbose = FALSE
)

kbas <- system.file("resources", "key_biodiversity_areas", "kbas.gpkg",
  package = "mapme.biodiversity")

aoi <- system.file("extdata", "shell_beach_protected_area_41057_B.gpkg",
  package = "mapme.biodiversity"
) \%>\%
  read_sf() \%>\%
  get_resources(get_key_biodiversity_area(kbas)) \%>\%
  calc_indicators(calc_key_biodiversity_area()) \%>\%
  portfolio_long()

aoi
}
}
\references{
BirdLife International (2024). The World Database of
Key Biodiversity Areas. Developed by the KBA Partnership: BirdLife
International, International Union for the Conservation of Nature,
Amphibian Survival Alliance, Conservation International, Critical Ecosystem
Partnership Fund, Global Environment Facility, Re:wild, NatureServe,
Rainforest Trust, Royal Society for the Protection of Birds, Wildlife
Conservation Society and World Wildlife Fund. Available at
www.keybiodiversityareas.org.
}
\keyword{indicator}
\keyword{resource}
