% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key_biodiversity_areas.R
\name{key_biodiversity_areas_resource}
\alias{key_biodiversity_areas_resource}
\alias{get_key_biodiversity_areas}
\title{Key Biodiversity Areas}
\source{
\url{https://www.keybiodiversityareas.org/kba-data}
}
\usage{
get_key_biodiversity_areas(path = NULL)
}
\arguments{
\item{path}{A character vector to the key biodiversity areas GPKG file.
Note, that the file has to be downloaded manually.}
}
\value{
A function that returns an \code{sf} footprints object.
}
\description{
This resource contains outlines of key biodiversity areas, which are areas
representing sites with specific importance for nature conservation.
}
\details{
To use this data in mapme workflows, you will have to manually download the
global data set and point towards its file path on your local machine.
Please find the available data under the source link given below.
}
\references{
BirdLife International (2024). The World Database of
Key Biodiversity Areas. Developed by the KBA Partnership: BirdLife
International, International Union for the Conservation of Nature,
Amphibian Survival Alliance, Conservation International, Critical Ecosystem
Partnership Fund, Global Environment Facility, Re:wild, NatureServe,
Rainforest Trust, Royal Society for the Protection of Birds, Wildlife
Conservation Society and World Wildlife Fund. Available at
www.keybiodiversityareas.org.
}
\keyword{resource}
