% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{make_footprints}
\alias{make_footprints}
\title{Create footprints for vector or raster data sets}
\usage{
make_footprints(files, what = c("vector", "raster"))
}
\arguments{
\item{files}{A character vector with GDAL readable paths to either vector
or raster sources.}

\item{what}{A character vector indicating if the files are vector or raster
files.}
}
\value{
An sf object with a the files sources and the geometry indicating
their spatial footprint.
}
\description{
With this function you can create footprints for vector or raster datasets.
Specify a character vector of GDAL readable files of either vector
or raster type. Internally, GDAL will be used to create an sf object
with a single column indicating the source and the geometry indicating
the bounding box of the respective file.
Note, the performance for remote files is dependent on your connection to
the server. If you have other means to create footprints in your resource
function (e.g. by using the output \code{{rstac::items_bbox()}}) you should prefer
those means over this function for remote files.
}
\examples{

# a vector resource
vec <- system.file("shape/nc.shp", package="sf")
make_footprints(vec, what = "vector")

# a raster resource
ras <- system.file("ex/elev.tif", package="terra")
make_footprints(ras, what = "raster")

}
