% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_global_grid}
\alias{make_global_grid}
\title{Helper to create a grid of regular resolution and CRS}
\usage{
make_global_grid(
  xmin = -180,
  xmax = 170,
  dx = 10,
  ymin = -50,
  ymax = 80,
  dy = 10,
  proj = NULL
)
}
\arguments{
\item{xmin}{minimum longitude value (E/W)}

\item{xmax}{maximum longitude value (E/W)}

\item{dx}{difference in longitude value per grid}

\item{ymin}{minimum latitude value (S/N)}

\item{ymax}{maximum latitude value (E/W)}

\item{dy}{difference in latitude value per grid}

\item{proj}{projection system}
}
\value{
An sf object with a defined grid.
}
\description{
Use this function to create a regular grid in a custom CRS. This is used
e.g. to create the tile grid for Global Forest Watch in order to retrieve
the intersecting tiles with a given portfolio.
}
\keyword{utils}
