% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapme.biodiversity-pkg.R
\docType{package}
\name{mapme.biodiversity - package}
\alias{mapme.biodiversity - package}
\title{mapme.biodiversity - Global biodiversity portfolio analysis}
\description{
mapme.biodiversity
}
\keyword{package}
