% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{mapme_options}
\alias{mapme_options}
\title{Options for mapme.biodiversity}
\usage{
mapme_options(..., outdir, verbose, aria_bin, testing)
}
\arguments{
\item{...}{ignored}

\item{outdir}{A length one character indicating the output path.}

\item{verbose}{A logical, indicating if informative messages should be printed.}

\item{aria_bin}{A character vector to an aria2c executable for parallel
downloads.}

\item{testing}{A logical. Not to be set by users. Controls the behavior
during automated test pipelines.}

\item{tempdir}{A length one character indicating the path to be used
for temporary files.}
}
\value{
A list of options if no arguments are specified. Otherwise sets
matching arguments to new values in the package's internal environment.
}
\description{
Sets default options for mapme.biodiversity to control the behavior of
downstream functions.
Mainly, the output path as well as the temporal directory for intermediate
files can be set. Additionally, the verbosity can be set. The testing options
should not be set by users, as it controls the behavior of the package during
automated test pipelines. Might be extended by other options in the future.
}
\examples{
library(mapme.biodiversity)
mapme_options()
}
