% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{mapme_portfolio}
\alias{mapme_portfolio}
\alias{init_portfolio}
\alias{as_portfolio}
\alias{as_portfolio.sf}
\alias{as_portfolio.character}
\alias{as_portfolio.tbl}
\alias{print.mapme_portfolio}
\alias{[.mapme_portfolio}
\alias{$<-.mapme_portfolio}
\alias{st_geometry.mapme_portfolio}
\title{Initiate a biodiversity portfolio}
\usage{
init_portfolio(x, years, sources)

as_portfolio(x, atts)

\method{as_portfolio}{sf}(x, atts)

\method{as_portfolio}{character}(x, atts)

\method{as_portfolio}{tbl}(x, atts)

\method{print}{mapme_portfolio}(x, ...)

\method{[}{mapme_portfolio}(x, i, j, drop = FALSE)

\method{$}{mapme_portfolio}(x, i) <- value

\method{st_geometry}{mapme_portfolio}(x, ...)
}
\arguments{
\item{x}{a mapme_portfolio}

\item{years}{A numeric vector indicating the years of analysis.}

\item{sources}{DEfaults to NA in case of a new initialization. Users can point
to a directory where resource have been previously downloaded.}

\item{atts}{All attributes of the original object + portfolio specific attributes}

\item{...}{additional arguments}

\item{i}{element index}

\item{j}{colomn indicator}

\item{drop}{logical from sf to make geometry column sticky}

\item{value}{the value(s) to assign to the element}
}
\value{
A \code{mapme_portfolio} object.
}
\description{
The function is to be used either on an \code{sf} object representing assets
for the initiation of a new portfolio or a character string pointing to a
Geopackage written to disk previously by \code{mapme.biodiversity}. Portfolio
wide parameters such as the timeframe of the analysis are to be added and initialization.

This function is an internal function used to frequently to convert
intermediate objects to a portfolio

Checks to be performed:
\itemize{
\item has at least one asset
\item all assets are either Polygon or MultiPolygon (TODO)
\item is Lat/Lon, otherwise transform to EPSG:4326
}

This function is used to re-read an exported Geopackage to the R Session
to allow users to continue their analysis in other sessions. Becuase the sizes
(e.g. number of rows/columns) very much differ from one indicator to another
in order to reduce file size geometries are stored in one table of the Geopackage
and metadata and indicator values are stored in related tables. This function
will re-rearange the input to the mapme_portfolio object on read time.

In order to support a selection of tidyverse verbs which return tibbles,
this function will create a mapme_portfolio object from these objects.
}
