% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean_species_abundance.R
\docType{data}
\name{mean_species_abundance_indicator}
\alias{mean_species_abundance_indicator}
\alias{calc_mean_species_abundance}
\title{Calculate Mean Species Abundance}
\format{
A function that returns an indicator tibble with variable
mean_species_abundance and corresponding values (unitless) as value.
}
\usage{
calc_mean_species_abundance()
}
\description{
This function calculates the average of the mean species abundance index for
a region.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{mean_species_abundance}
}
}
\examples{
\dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
}
\dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

mapme_options(
  outdir = outdir,
  verbose = FALSE
)

msa <- system.file("res", "mean_species_abundance",
                   "TerrestrialMSA_2015_World.tif",
                   package = "mapme.biodiversity")

aoi <- read_sf(
  system.file("extdata", "shell_beach_protected_area_41057_B.gpkg",
              package = "mapme.biodiversity"
))
aoi <- get_resources(aoi, get_mean_species_abundance(msa))
aoi <- calc_indicators(aoi, calc_mean_species_abundance())
aoi <- portfolio_long(aoi)

aoi
}
}
\keyword{indicator}
