% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_species_abundance.R
\name{mean_species_abundance_resource}
\alias{mean_species_abundance_resource}
\alias{get_mean_species_abundance}
\title{Mean Species Abundance}
\source{
\url{https://www.globio.info/globio-data-downloads}
}
\usage{
get_mean_species_abundance(path = NULL)
}
\arguments{
\item{path}{A character vector to the msa data file}
}
\value{
A function that returns an \code{sf} footprints object.
}
\description{
The variable represents the mean species abundance (MSA), for plants and
warmblooded vertabrates: values ranging from 0 to 1 indicating local
biodiversity intactness, relative to a pristine reference situation.
}
\details{
To use this data in mapme workflows, you will have to manually download the
global data set and point towards the file path on your local machine.
Please find the available data under the source link given below.
}
\references{
Schipper, AM, Hilbers, JP, Meijer, JR, et al. Projecting
terrestrial biodiversity intactness with GLOBIO 4. Glob Change Biol. 2020;
26: 760– 771.
\doi{https://doi.org/10.1111/gcb.14848}
}
\keyword{resource}
