% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa_grace.R
\name{nasa_grace}
\alias{nasa_grace}
\alias{get_nasa_grace}
\title{NASA GRACE-based Drought Indicator layer}
\usage{
get_nasa_grace(years = 2003:2022)
}
\arguments{
\item{years}{A numeric vector indicating the years for which to make the
resource available.}
}
\value{
A function that returns a character of file paths.
}
\description{
The resource is published by NASA GRACE Tellus. This data set
reflects on potential drought conditions in the shallow groundwater section
relative to a reference period spanning from 1948 to 2012.
It is available as a global raster with a weekly temporal resolution starting
with the year 2003. The value indicates the wetness percentile of a given
pixel with regard to the reference period.
}
\keyword{resource}
