% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_droughtInd.R
\docType{data}
\name{nasagrace}
\alias{nasagrace}
\title{NASA GRACE-based Drought Indicator layer}
\format{
Global raster layers available for years 2003 to present.
}
\source{
\url{https://nasagrace.unl.edu/globaldata/}
}
\description{
This resource is published by NASA GRACE Tellus. It indicates the drought,
global raster layer available from year 2003 to present with weekly updates.
Encoded as percentile, representing current wet or dry conditions showing the
probability of occurrence - with lower values meaning dryer than normal, and
higher values meaning wetter than normal.
}
\keyword{resource}
