% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ghsl_indicators.R
\name{population_count_ghsl}
\alias{population_count_ghsl}
\alias{calc_population_count_ghsl}
\title{Calculate population count statistics based on GHSL}
\usage{
calc_population_count_ghsl(engine = "extract", stats = "sum")
}
\arguments{
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}

\item{stats}{Function to be applied to compute statistics for polygons
either one or multiple inputs as character "min", "max", "sum", "mean", "median"
"sd" or "var".}
}
\value{
A function that returns an indicator tibble with the specified
populations statistics as variable and the corresponding values as value.
}
\description{
The Global Human Settlement Layer (GHSL) provides a set of global spatial
information, evidence-based analytics, and knowledge about the human
presence on the planet. This function allows to efficiently
calculate population count statistics (e.g. total number of population) for
polygons based on the GHS-POP dataset. For each polygon, the desired
statistic/s (min, max, sum, mean, median, sd or var) is/are returned.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{ghsl}
}
}
\keyword{indicator}
