% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_popCount.R
\name{PopulationCount_Global_Mosaics}
\alias{PopulationCount_Global_Mosaics}
\alias{.get_popCount}
\title{Downloads Population Count layer}
\usage{
.get_popCount(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
This resource represents the population count, layers available to download from the
year 2000 to 2020. The dataset is called as WorldPop Unconstrained Global Mosaics.
The encoded cell value represents the total number of people in that particular cell.
}
\keyword{internal}
