% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_precipitation.R
\docType{data}
\name{precipitation}
\alias{precipitation}
\title{Calculate precipitation statistics}
\format{
A tibble with a column for years, months, the statistic and the coressponding value (in mm)
}
\description{
This functions allows to calculate precipitation statistics based on the
CHIRPS rainfall estimates. Corresponding to the time-frame of the analysis
of the portfolio, monthly precipitation statistics are calculated. Users
can decide which statistics to calculate.
The required resources for this indicator are:
\itemize{
\item \code{chirps}
}
}
\details{
The following arguments can be set:
\describe{
\item{stats}{The statistic to calculate. One or multiple of: 'mean', 'median', 'sd'.}
}
}
\keyword{indicator}
