% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_proximity.R
\docType{data}
\name{proximity_wdpa}
\alias{proximity_wdpa}
\alias{calc_proximity_wdpa}
\title{Calculate Proximity to Protected Areas from WDPA}
\format{
A function returning an indicator tibble with \code{proximity_wdpa}
as variable and the minimal distance (in meters) as value.
}
\usage{
calc_proximity_wdpa(asset_column = NULL)
}
\description{
This function calculates the distance to the closest WDPA within an area
of influence around an assets location.
}
\details{
The required resources for this indicator are:
\itemize{
\item \link{wdpa}
}

Note, that for this indicator to function properly, two distinct geometry
columns in the portfolio are required. The active geometry column must
represent the area of influence, as the KBA resource will be clipped to
its extent during read time. A second inactive geometry columns must
be specified per the 'asset_column' parameter. It must not be equal
to the active geometry column.

See the example code on how to calculate and activating a second geometry
column.
}
\examples{
\dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
}
\dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

mapme_options(
  outdir = outdir,
  chunk_size = 1e8,
  verbose = FALSE
)

aoi <- read_sf(
  system.file("extdata", "shell_beach_protected_area_41057_B.gpkg",
    package = "mapme.biodiversity"
  )
)

asset_column <- attr(aoi, "sf_column")

aoi$buffer <- st_geometry(st_buffer(aoi, dist = 50000))
st_geometry(aoi) <- "buffer"

aoi <- get_resources(aoi, get_wdpa(version = "Oct2024"))
aoi <- calc_indicators(aoi, calc_proximity_wdpa(asset_column))

aoi$buffer <- NULL
st_geometry(aoi) <- asset_column

aoi <- portfolio_long(aoi)

aoi
}
}
\keyword{indicator}
