% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_indicator}
\alias{register_indicator}
\title{Register a new indicator to mapme.biodiversity}
\usage{
register_indicator(name = NULL, description = NULL, resources = NULL)
}
\arguments{
\item{name}{A character vector indicating the name of the indicator.}

\item{description}{A character vector with a basic description of
the indicator}

\item{resources}{A character vector of the required resources
that need to be available to calculate the indicator. The names must
correspond with already registered resources.}
}
\value{
Nothing. Registers the function in the package environment.
}
\description{
Registers a custom indicator function to access indicators not native to
\code{mapme.biodiversity} to be used with \code{calc_indicators()}. Custom
indicators will also be registered in the list generated by
\code{available_indicators()}.
}
\details{
Note that registering your own indicator function will only have effect for
the current R session. If you return to your analysis in a new session, you
will have to re-register your custom indicator.
}
\examples{
\dontrun{
register_indicator(
  name = "treecover_area",
  description = "Area of forest cover by year",
  resources = c(
    "gfw_treecover",
    "gfw_lossyear"
  )
)
}
}
\keyword{utils}
