% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{select_engine}
\alias{select_engine}
\title{Extract values with an engine and statistics}
\usage{
select_engine(x, raster, stats, engine, name = NULL, mode = "asset")
}
\arguments{
\item{x}{An sf object representing a portfolio.}

\item{raster}{An terra SpatRaster from which values are to be extracted.}

\item{stats}{A character vector of statistics to aggregate the raster values
with.}

\item{engine}{A character vector of length one specifying the engine to be
used for the extraction.}

\item{name}{A character vector indicating the name to append to the columns
names.}

\item{mode}{A character vector indicating in which mode to conduct the
extraction (e.g. \code{asset}-wise or for the whole \code{portfolio} at once).}
}
\value{
A tibble.
}
\description{
Use this function for zonal vector-raster value extractions based on
a supported engine and for one or more statistics. Columns are named
according to the argument \code{name} plus the respective stat. Both \code{portfolio}
and \code{asset} mode are supported.
}
\keyword{utils}
