% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SRTMdem.R
\name{SRTM_Digital_Elevation_Model}
\alias{SRTM_Digital_Elevation_Model}
\alias{.get_SRTMdem}
\title{Downloads SRTM 30m Digital Elevation Model (DEM) Layer}
\usage{
.get_SRTMdem(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
The layer represents the 30m global terrestrial digital elevation model from
the NASA Shuttle Radar Topographic Mission (SRTM), available for download as
5 degree x 5 degree tiles. It is encoded as meter, representing the elevation
at the particular grid cell.
}
\keyword{internal}
