% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_star.R
\name{star}
\alias{star}
\alias{get_star}
\title{Species Threat Abatement and Restoration (STAR) layers}
\source{
\url{https://www.nature.com/articles/s41559-021-01432-0}
}
\usage{
get_star(layer = c("both", "abatement", "restoration"))
}
\arguments{
\item{paths}{A character vector to the respective species range files in GTiff
format. Note, that theses files have to be downloaded manually.}
}
\value{
A function that returns an \code{sf} footprint object.
}
\description{
STAR layers are calculated based on the current and hypothetical area of
habitat (AOH) for species considered under threat by the IUCN Red List.
The version used here is the 50 km resolution published with the original
research paper. Thus it should only be used for continental/global scale
analysis.
}
\details{
The threat abatement layer is calculated for each resolution cell by considering
the current AOH for each species weighted by its threat level (Near Threatened = 1;
Vulnerable = 2; Endangered = 3; Critically Endangered = 4) and the
relative contribution of different thread categories on that species.

The restoration layer considers the extent of restorable AOH at each resolution
cell weighted by species threat level (see above), threat category, and
additionally weighted by a global multiplier to discount restoration scores.

Consider the methods section of the original publication for additional
information.
}
\references{
Mair, L., Bennun, L.A., Brooks, T.M. et al. A metric for
spatially explicit contributions to science-based species targets.
Nat Ecol Evol 5, 836–844 (2021). \doi{https://doi.org/10.1038/s41559-021-01432-0}
}
\keyword{resource}
