% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_star.R
\name{star_stats}
\alias{star_stats}
\alias{calc_star}
\title{Species threat abatement and restoration (STAR) zonal statistics}
\usage{
calc_star(engine = "extract", stats = "mean")
}
\arguments{
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}

\item{stats}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character. Supported statistics are: "mean",
"median", "sd", "min", "max", "sum" "var".}
}
\value{
A function that returns an indicator tibble with IUCN layers with
specified statistics as variable and respective species richness (count)
as value.
}
\description{
Calculates zonal statistics of the STAR resources raster layers.
}
\details{
For threat abatement, low values (towards 0) represent locations with
low levels of threat for all present species, while higher values (>5000),
represent areas with large number of species under significant threats.

For restoration, lower values (towards 0) represent areas where restoration efforts are
less likely to contribute to lowering the threat level on the present species,
while higher values (>5000) indicate areas where restoration efforts are
very likley to result in lowering the threat levels of the present species.

Be referred to the documentation of the STAR resource and the original
research article for details on interpreting its values.

The required resources for this indicator are:
\itemize{
\item \link{star}
}
}
\keyword{indicator}
