% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{[[<-.mapme_portfolio}
\alias{[[<-.mapme_portfolio}
\title{Dobuble bracket assigmnent method for mapme_portfolios}
\usage{
\method{[[}{mapme_portfolio}(x, i) <- value
}
\arguments{
\item{x}{a mapme_portfolio object}

\item{i}{element index}

\item{value}{the value(s) to assign to the element}
}
\description{
Dobuble bracket assigmnent method for mapme_portfolios
}
