% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecoregions.R
\name{Terrestrial_Ecoregions_of_the_World}
\alias{Terrestrial_Ecoregions_of_the_World}
\alias{.get_ecoregions}
\title{Downloads Terrestrial Ecoregions of the World (TEOW) Polygon}
\usage{
.get_ecoregions(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
The polygons represent the ecoregions, defined as relatively large units of land
or inland water sharing a large majority of biodiversity. The datasets is made
available from World Wildlife Fund (WWF). There are 867 terrestrial ecoregions,
further classified into 14 different terrestrial biomes.
}
\keyword{internal}
