% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{tidyverse}
\alias{tidyverse}
\alias{group_by.mapme_portfolio}
\alias{group_split.mapme_portfolio}
\title{Support for group_by for mapme_portfolios}
\usage{
\method{group_by}{mapme_portfolio}(.data, ...)

\method{group_split}{mapme_portfolio}(.data, ...)
}
\arguments{
\item{.data}{data object of class \link{mapme_portfolio}}

\item{...}{other arguments}
}
\description{
Support for group_by for mapme_portfolios

Support for group_split for mapme_portfilios
}
