% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accessibility.R
\name{Travel_time_to_major_Cities}
\alias{Travel_time_to_major_Cities}
\alias{.get_accessibility}
\title{Downloads Accessibility to Cities layer}
\usage{
.get_accessibility(
  x,
  range_accessibility = "20k_50k",
  rundir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{range_accessibility}{The city within the defined range of population to download, defaults to \code{"20k_50k"}.}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
Accessibility is the ease with which larger cities can be reached from a certain location.
This resource represents the travel time to major cities in the year 2015. Encoded as minutes,
representing the time needed to reach that particular cell from nearby city of population range
of interest.
}
\keyword{internal}
