% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_accessibility.R
\docType{data}
\name{traveltime}
\alias{traveltime}
\title{Accessibility to Cities layer}
\format{
Global raster layer available for year 2015.
}
\source{
\url{https://figshare.com/articles/dataset/Travel_time_to_cities_and_ports_in_the_year_2015/7638134/3}
}
\description{
This resource is published by Weiss et al. (2018) "A global map of travel
time to cities to assess inequalities in accessibility in 2015" on journal
nature. Accessibility is the ease with which larger cities can be reached
from a certain location. This resource represents the travel time to major
cities in the year 2015. Encoded as minutes, representing the time needed
to reach that particular cell from nearby city of target population range.
}
\references{
Weiss, D. J., Nelson, A., Gibson, H. S., Temperley, W., Peedell,
S., Lieber, A., … & Gething, P. W. (2018). A global map of travel time to cities
to assess inequalities in accessibility in 2015. Nature, 553(7688), 333-336.
}
\keyword{resource}
