% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_treecover.R
\docType{data}
\name{treecover}
\alias{treecover}
\title{Calculate treecover statistics}
\format{
A tibble with a column for years and treecover (in ha)
}
\description{
This functions allows to efficiently calculate treecover statistics for
polygons. For each year in the analysis timeframe, the forest losses in
preceeding and the current years are substracted from the treecover in the
year 2000 and actual treecover figures within the polygon are returned.
The required resources for this indicator are:
\itemize{
\item \link{treecover2000}
\item \link{lossyear}
}
}
\details{
The following arguments can be set:
\describe{
\item{min_size}{The minum size of a forest patch to be considered as forest in ha.}
\item{min_cover}{The minimum cover percentage per pixel to be considered as forest.}
}
}
\keyword{indicator}
