% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ucdp_ged.R
\docType{data}
\name{ucdp_ged}
\alias{ucdp_ged}
\title{UCDP Georeferenced Event Dataset (UCDP GED)}
\format{
A global event dataset (GED) encoding envents of organized
violence a point geometries
}
\source{
\url{https://ucdp.uu.se/downloads/}
}
\description{
This resource distributed by the Uppsala Conflict Data Program (UCDP) constitutes
its most diaggregated dataset on individual events of organized violence.
It encodes the different actors involved, is spatially disaggregated down
to village levels anc currently covers the time period of 1989 to 2021.
Older versions of the data set can be downloaded, but users are recommended
to download the latest data set.
}
\details{
The following versions are available
\itemize{
\item 5.0
\item 17.1
\item 17.2
\item 18.1
\item 19.1
\item 20.1
\item 21.1
\item 22.1 or latest
}

The following argument should be specified by users:

\describe{
\item{version_ged}{A character vector specifying
the version to download. Defaults to "latest".}
}
}
\references{
Davies, Shawn, Therese Pettersson & Magnus Öberg (2022).
Organized violence 1989-2021 and drone warfare.
Journal of Peace Research 59(4).
\doi{10.1177/00223433221108428}
}
\keyword{resource}
