% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wdpa.R
\name{wdpa}
\alias{wdpa}
\alias{get_wdpa}
\title{World Database on Protected Areas}
\source{
\url{https://www.protectedplanet.net/}
}
\usage{
get_wdpa(version = "latest")
}
\arguments{
\item{version}{A character vector indicating the version to be made available.
Must contain the month as a three letter combination and the year, e.g. "Aug2024",
or "latest" (the default) meaning the data from the previous month.}
}
\value{
A function that returns an \code{sf} footprints object.
}
\description{
This resource contains outlines of protected areas around the globe
which are part of the World Database on Protected Areas. This database
is compiled by UN Environment Programme World Conservation Monitoring Centre
(UNEP-WCMC), in collaboration with governments, non-governmental
organisations, academia and industry.
}
\details{
The data is updated with a monthly cadence. Here, we only retain protected
areas with polygon information available, i.e. areas for which only
a point location is reported are omitted. Additionally, we apply
a function to repair geometries from GEOS wrap polygons that cross the
dateline.
}
\references{
UNEP-WCMC and IUCN (2024), Protected Planet: The World Database
on Protected Areas (WDPA) 2024, Cambridge, UK: UNEP-WCMC and IUCN.
Available at: www.protectedplanet.net.
}
\keyword{resource}
