% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_worldclim.R
\docType{data}
\name{worldclim}
\alias{worldclim}
\title{WorldClim climatic variables (min temperature, max temperature, mean precipitation)}
\format{
Global raster layers available for years 2000 to 2018.
}
\source{
\url{https://www.worldclim.org/data/index.html}
}
\description{
This resource is published by Fick et al. (2017) "WorldClim 2: new 1-km
spatial resolution climate surfaces for global land areas". This resource
represents multiple climatic variables from which we will be requiring minimum
temperature, maximum temperature, and mean precipitation layers. The layers are
available to download for the period 2000 - 2018 on monthly basis from WorldClim.
}
\details{
Enlisted different resources can be requested with their dedicated functions:
\describe{
\item{tmin}{Encoded as (°C), representing the minimum temperature per output grid cell.}
\item{tmax}{Encoded as (°C), representing the maximum temperature per output grid cell.}
\item{prec}{Encoded as (mm), representing the mean precipitation per output grid cell.}
}
}
\keyword{resource}
