% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climaticVariables.R
\name{WorldClim_Mean_Precipitation}
\alias{WorldClim_Mean_Precipitation}
\alias{.get_precipitation}
\title{Downloads WorldClim Mean Precipitation layer}
\usage{
.get_precipitation(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
This resource represents the average precipitation, layers available to download for
the period 2000 - 2018 on monthly basis from WorldClim. Encoded as (mm), representing
the mean precipitation per output grid cell.
}
\keyword{internal}
