% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_worldclim_mintemperature.R
\docType{data}
\name{worldclim_tmin}
\alias{worldclim_tmin}
\title{Calculate minimum temperature statistics}
\format{
A tibble with a column for minimum temperature statistics (in °C)
}
\description{
This function allows to efficiently calculate minimum temperature statistics
from Worldclim for polygons. For each polygon, the desired statistic/s (min,
max, sum, mean, median, sd or var) is/are returned.
The required resources for this indicator are:
\itemize{
\item \code{mintemperature}
}
}
\details{
The following arguments can be set:
\describe{
\item{stats_mintemperature}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character. Supported statistics are: "mean",
"median", "sd", "min", "max", "sum" "var".}
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}
}
}
\keyword{indicator}
