% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_worldpop.R
\docType{data}
\name{worldpop}
\alias{worldpop}
\title{Population Count layer for year 2000-2020}
\format{
Global raster layers available for years 2000-2020.
}
\source{
\url{https://www.worldpop.org/geodata/listing?id=64}
}
\description{
This resource is published by open spatial demographic data and research
organization called WorldPop. This resource represents the population
count, 1 km spatial resolution layers available to download from the year
2000 to 2020. The dataset is called as WorldPop Unconstrained Global Mosaics.
The encoded cell value represents the total number of people in that particular
grid cell.
}
\keyword{resource}
