% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lossyear.R
\name{Year_of_Gross_Forest_Loss}
\alias{Year_of_Gross_Forest_Loss}
\alias{.get_lossyear}
\title{Downloads Year of Gross Forest Loss Event}
\usage{
.get_lossyear(
  x,
  vers_lossyear = "GFC-2020-v1.8",
  rundir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{vers_lossyear}{The version to download, defaults to
\code{"GFC-2020-v1.8"}.}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{A directory where intermediate files are written to.}
}
\description{
The layer is defined as forest loss during the period 2000–2019, defined as
a stand-replacement disturbance, or a change from a forest to non-forest
state. Encoded as either 0 (no loss) or else a value in the range 1–19,
representing loss detected primarily in the year 2001–2019, respectively.
}
\keyword{internal}
