% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{available_indicators}
\alias{available_indicators}
\title{Backlog function for available indicators}
\usage{
available_indicators(indicators = NULL)
}
\arguments{
\item{indicators}{If \code{NULL} returns a list of all registered indicators
(default). Otherwise only the ones specified.}
}
\value{
A list object.
}
\description{
Returns a list of indicator names and parametrization options for the
specified indicators. If no resource names are provided, it lists all
available indicators, including custom registered ones.
Use it to learn about possible additional arguments that can be specified
when requesting an indicator.
}
\examples{
names(available_indicators())
}
\keyword{indicator}
