% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_namespace}
\alias{check_namespace}
\title{Checks if namespace is available}
\usage{
check_namespace(pkg)
}
\arguments{
\item{pkg}{A character vector of length one indicating a package name
for which the namespace is tested}
}
\value{
TRUE, invisible, if the namespace is available. An error message
otherwise.
}
\description{
Use this function if your resource/indicator function requires the
namespace of a certain package to be available. An informative error
message is printed if that is not the case.
}
\keyword{utils}
