% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dem.R
\name{.calc_dem}
\alias{.calc_dem}
\title{Calculate elevation statistics based on SRTM data sets}
\usage{
.calc_dem(
  shp,
  srtmdem,
  engine = "zonal",
  stats_elevation = "mean",
  rundir = tempdir(),
  verbose = TRUE,
  todisk = FALSE,
  ...
)
}
\arguments{
\item{shp}{A single polygon for which to calculate the elevation statistic}

\item{srtmdem}{The elevation raster resource from SRTM}

\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{A directory where intermediate files are written to.}

\item{todisk}{Logical indicating whether or not temporary raster files shall
be written to disk}

\item{...}{additional arguments}

\item{stats}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character "mean", "median" or "sd".}
}
\value{
A tibble
}
\description{
Considering the 30m resolution SRTM raster datasets users can specify which
statistics among mean, median or standard deviation to compute. Also, users
can specify the functions i.e. zonal from package terra, extract from package
terra, or exactextract from exactextractr as desired.
}
\keyword{internal}
