% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tri.R
\name{.calc_tri}
\alias{.calc_tri}
\title{Calculate Terrain Ruggedness Index (TRI) based on SRTM rasters}
\usage{
.calc_tri(
  shp,
  srtmelevation,
  engine = "zonal",
  stats = "mean",
  rundir = tempdir(),
  verbose = TRUE,
  todisk = FALSE,
  ...
)
}
\description{
Calculate Terrain Ruggedness Index (TRI) based on SRTM rasters
}
\keyword{internal}
