% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tri.R
\name{.comp_tri_exact_extractr}
\alias{.comp_tri_exact_extractr}
\title{function to compute single raster with exactextractr}
\usage{
.comp_tri_exact_extractr(
  elevation = NULL,
  shp = NULL,
  stats = "mean",
  todisk = todisk,
  rundir = rundir,
  ...
)
}
\description{
function to compute single raster with exactextractr
}
\keyword{internal}
