% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tri.R
\name{.comp_tri_zonal}
\alias{.comp_tri_zonal}
\title{function to compute single raster with zonal}
\usage{
.comp_tri_zonal(
  elevation = NULL,
  shp = NULL,
  stats = "mean",
  todisk = FALSE,
  rundir = tempdir,
  ...
)
}
\description{
function to compute single raster with zonal
}
\keyword{internal}
