% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mangrove.R
\name{.unzipMangrove}
\alias{.unzipMangrove}
\title{Helper function to correctly unzip a mangrove layer zip file}
\usage{
.unzipMangrove(zip, rundir)
}
\arguments{
\item{rundir}{The directory to where the files are unzipped}

\item{zip_files}{A charchter vector with potentially multiple zip files}
}
\value{
Nothing.
}
\description{
Helper function to correctly unzip a mangrove layer zip file
}
\keyword{internal}
