% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_droughtInd.R
\name{Drought_Indicator}
\alias{Drought_Indicator}
\alias{.get_drought_ind}
\title{Downloads NASA GRACE-based Drought Indicator Layer}
\usage{
.get_drought_ind(x, rundir = tempdir(), verbose = TRUE)
}
\arguments{
\item{x}{An sf object returned by init_portfolio}

\item{rundir}{A directory where intermediate files are written to.}

\item{verbose}{Logical controlling verbosity.}
}
\description{
This resource indicates the drought, global raster layer available from year
2003 to present with weekly updates. Encoded as percentile, representing
current wet or dry conditions showing the probability of occurrence - with
lower values meaning dryer than normal, and higher values meaning wetter than
normal.
}
\keyword{internal}
