% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dem.R
\docType{data}
\name{elevation}
\alias{elevation}
\title{Calculate elevation statistics}
\format{
A tibble with a column for elevation statistics (in meters)
}
\description{
This function allows to efficiently calculate elevation statistics for
polygons. For each polygon, the desired statistic/s (mean, median or sd)
is/are returned.
The required resources for this indicator are:
\itemize{
\item \code{srtmdem}
}
}
\details{
The following arguments can be set:
\describe{
\item{stats}{Function to be applied to compute statistics for polygons either
one or multiple inputs as character "mean", "median" or "sd".}
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}
}
}
\keyword{indicator}
