% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_esalandcover.R
\docType{data}
\name{esalandcover}
\alias{esalandcover}
\title{ESA Copernicus Global Land Cover layer}
\format{
A global tiled raster resource available for years 2015 to 2019.
}
\source{
\url{https://lcviewer.vito.be/download}
}
\description{
This 100 meter spatial resolution land cover resource is published by Buchhorn
et al. (2020) "Copernicus Global Land Cover Layers—Collection 2". The resource
represents the actual surface cover of ground available annually for the period
2015 to 2019. The cell values range from 0 to 200, representing total of 23
discrete classifications from ESA.
}
\references{
© European Union, Copernicus Land Monitoring Service (year),
European Environment Agency (EEA)", f.ex. in 2018: “© European Union,
Copernicus Land Monitoring Service 2018, European Environment Agency (EEA)
}
\keyword{resource}
