% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gsw.R
\docType{data}
\name{global_surface_water_occurrence}
\alias{global_surface_water_occurrence}
\title{Global Surface Water Occurrence}
\format{
Global raster layers
}
\source{
\url{https://global-surface-water.appspot.com/}
}
\description{
GSW occurrence raw data comes in raster files with integer cell values
between \code{[0, 100]}. This value gives the percentage of the time that a
given pixel was classified as water during the entire observation period. So
a 0 denotes a pixel that was never classified as water, 100 denotes a pixel
with permanent water.
}
\keyword{resource}
