% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gmw.R
\docType{data}
\name{gmw}
\alias{gmw}
\title{Global Mangrove Extent Polygon}
\format{
Global mangrove extent polygon available for years 1996, 2007-2010,
2015, and 2016.
}
\source{
\url{https://data.unep-wcmc.org/datasets/45}
}
\description{
This resource is part of the publication by Bunting et al. (2018)
"The Global Mangrove Watch—A New 2010 Global Baseline of Mangrove Extent".
The polygons represent the mangrove, which is tropical coastal vegetation
and considered the most significant part of the marine ecosystem. This
resource is available for the period 1996- 2016 from Global Mangrove Watch
(GMW), providing geospatial information about global mangrove extent.
}
\references{
Bunting P., Rosenqvist A., Lucas R., Rebelo L-M., Hilarides L.,
Thomas N., Hardy A., Itoh T., Shimada M. and Finlayson C.M. (2018). The Global
Mangrove Watch – a New 2010 Global Baseline of Mangrove Extent. Remote Sensing
10(10): 1669. doi:10.3390/rs10101669.
}
\keyword{resource}
