% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_gsw_recurrence.R
\docType{data}
\name{gsw_recurrence}
\alias{gsw_recurrence}
\title{Calculate Global Surface Water (GSW) Recurrence}
\format{
A tibble with a column for the aggregated GSW recurrence indicator.
}
\description{
Water Recurrence is a measurement of the degree of variability in the
presence of water from year to year. It describes the frequency with which
water returned to a particular location from one year to another, and is
expressed as a percentage. The raster files have integer cell values between
\code{[0, 100]}, where 100 represents that water reoccurs predictably every
year, whereas lower values indicate that water only occurs episodically.
}
\details{
The pixel values are aggregated using method provided via the
\code{stats_gsw} parameter.

The required resources for this indicator are:
\itemize{
\item \link{global_surface_water_recurrence}
}

The following arguments can be set:
\describe{
\item{stats_gsw}{The aggregation function applied to the input raster
values. Defaults to \code{mean}.}
}
}
\examples{
\dontshow{
mapme.biodiversity:::.copy_resource_dir(file.path(tempdir(), "mapme-data"))
}
\dontrun{
library(sf)
library(mapme.biodiversity)

outdir <- file.path(tempdir(), "mapme-data")
dir.create(outdir, showWarnings = FALSE)

aoi <- system.file("extdata", "shell_beach_protected_area_41057_B.gpkg",
  package = "mapme.biodiversity"
) \%>\%
  read_sf() \%>\%
  init_portfolio(
    years = 2001,
    outdir = outdir,
    tmpdir = tempdir(),
    verbose = FALSE
  ) \%>\%
  get_resources("global_surface_water_recurrence") \%>\%
  calc_indicators("gsw_recurrence")

aoi
}
}
\keyword{indicator}
