% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa_srtm.R
\name{nasa_srtm}
\alias{nasa_srtm}
\alias{get_nasa_srtm}
\title{NASADEM HGT v001}
\source{
https://planetarycomputer.microsoft.com/dataset/nasadem
}
\usage{
get_nasa_srtm()
}
\value{
A function that returns a character of file paths.
}
\description{
This resource is processed by the Land Processes Distributed Active Archive
Center (LP DAAC) and made available at the Microsoft Planetery Computer.
NASADEM are distributed in 1 degree latitude by 1 degree longitude tiles and
consist of all land between 60° N and 56° S latitude. This accounts for about
80\% of Earth’s total landmass.
}
\references{
NASA JPL (2020). NASADEM Merged DEM Global 1 arc second V001.
NASA EOSDIS Land Processes DAAC. Accessed 2023-07-01 from
https://doi.org/10.5067/MEaSUREs/NASADEM/NASADEM_HGT.001
}
\keyword{resource}
