% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasa_srtm.R
\docType{data}
\name{nasa_srtm}
\alias{nasa_srtm}
\title{SRTM 30m Digital Elevation Model (DEM) layer}
\format{
A global tiled raster resource available for all land areas.
}
\source{
\url{https://srtm.csi.cgiar.org/}
}
\description{
This resource is published by Farr et al. (2007) "The Shuttle Radar Topography
Mission". The layer represents the 30m global terrestrial digital elevation model
from the NASA Shuttle Radar Topographic Mission (SRTM), available for download as
5 degree x 5 degree tiles. It is encoded as meter, representing the elevation
at the particular grid cell.
}
\references{
NASA Shuttle Radar Topography Mission (SRTM)(2013). Shuttle Radar
Topography Mission (SRTM) Global. Distributed by OpenTopography.
https://doi.org/10.5069/G9445JDF. Accessed: 2022-03-17
}
\keyword{resource}
