% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nasagrace.R
\docType{data}
\name{nasagrace}
\alias{nasagrace}
\title{NASA GRACE-based Drought Indicator layer}
\format{
Global raster layers available for years 2003 to present.
}
\source{
\url{https://nasagrace.unl.edu/globaldata/}
}
\description{
The resource is published by NASA GRACE Tellus. This data set
reflects on potential drought conditions in the shallow groundwater section
relative to a reference period spanning from 1948 to 2012.
It is available as a global raster with a weekly temporal resolution starting
with the year 2003. The value indicates the wetness percentile of a given
pixel with regard to the reference period.
}
\keyword{resource}
