% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_precipitation.R
\docType{data}
\name{precipitation}
\alias{precipitation}
\title{Calculate precipitation statistics}
\format{
A tibble with a column for years, months, absolute rainfall, rainfall
anomaly and one or more columns per selected time-scale for SPI.
}
\description{
This functions allows to calculate precipitation statistics based on the
CHIRPS rainfall estimates. Corresponding to the time-frame of the analysis
of the portfolio, monthly precipitation statistics are calculated. These include
the total rainfall amount, rainfall anomaly against the 1981-2010 climate normal,
and the Standardized Precipitation Index (SPI) which is available for scales
between 1 and 48 months. Th function needs the \code{SPEI} package to be
installed.
The required resources for this indicator are:
\itemize{
\item \code{chirps}
}
}
\details{
The following arguments can be set:
\describe{
\item{scales_spi}{An integer vector indicating the scales for which to calculate the SPI.}
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}
}
}
\keyword{indicator}
