% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_indicator}
\alias{register_indicator}
\title{Register a new indicator to mapme.biodiversity}
\usage{
register_indicator(name = NULL, resources = NULL)
}
\arguments{
\item{name}{A character vector indicating the name of the indicator.}

\item{resources}{A list with named objects indicating the resources
that need to be available to calculate the indicator. The names correspond
to registered resources and a single character value indicates the
type of that resources}
}
\value{
Nothing. Registers the function in the package environment.
}
\description{
Registers a custom indicator function to access indicators not native to
\code{mapme.biodiversity} to be used with \code{calc_indicators()}. Custom
indicators will also be registered in the list generated by
\code{available_indicators()}.
}
\details{
Note that registering your own indicator function will only have effect for
the current R session. If you return to your analysis in a new session, you
will have to re-register your custom indicator.
}
\examples{
\dontrun{
register_indicator(
  name = "treecover_area",
  resources = list(
    gfw_treecover = "raster",
    gfw_lossyear = "raster"
  )
)
}
}
\keyword{utils}
