% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teow.R
\name{teow}
\alias{teow}
\alias{get_teow}
\title{Terrestrial Ecoregions of the World (TEOW) Polygon}
\usage{
get_teow()
}
\value{
A function that returns an \code{sf} footprint object.
}
\description{
This resource is part of the publication by Olson et al. (2004)
"Terrestrial Ecosystems of the World (TEOW) from WWF-US (Olson)". It
depicts 867 terrestrial ecoregions around the world classified into 14
different terrestrial biomes such as forests, grasslands, or deserts.
The polygons represent the ecoregions, defined as relatively large units of
land or inland water sharing a large majority of biodiversity. The datasets
is made available from World Wildlife Fund (WWF) for the year 2001.
}
\references{
Olson, D. M., Dinerstein, E., Wikramanayake, E. D., Burgess, N. D.,
Powell, G. V. N., Underwood, E. C., D’Amico, J. A., Itoua, I., Strand, H. E.,
Morrison, J. C., Loucks, C. J., Allnutt, T. F., Ricketts, T. H., Kura, Y., Lamoreux,
J. F., Wettengel, W. W., Hedao, P., Kassem, K. R. 2001. Terrestrial ecoregions of
the world: a new map of life on Earth. Bioscience 51(11):933-938.
\doi{https://doi.org/10.1641/0006-3568(2001)051[0933:TEOTWA]2.0.CO;2}
}
\keyword{resource}
