% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tri.R
\docType{data}
\name{tri}
\alias{tri}
\title{Calculate Terrain Ruggedness Index (TRI) statistics}
\format{
A tibble with a column for terrain ruggedness index statistics (in meters).
The range of index values and corresponding meaning:
(1) 0 - 80 m  :- level surface
(2) 81-116 m  :- nearly level surface
(3) 117-161 m :- slightly rugged surface
(4) 162-239 m :- intermediately rugged surface
(5) 240-497 m :- moderately rugged surface
(6) 498-958 m :- highly rugged surface
(7) 959-4367 m:- extremely rugged surface
}
\description{
Terrain Ruggedness Index is a measurement developed by Riley, et al. (1999).
The elevation difference between the center pixel and its eight immediate
pixels are squared and then averaged and its square root is taken to get
the TRI value. This function allows to efficiently calculate terrain ruggedness
index (tri) statistics for polygons. For each polygon, the desired statistic/s
(mean, median or sd) is/are returned.
The required resources for this indicator are:
\itemize{
\item \link{srtmdem}
}
}
\details{
The following arguments can be set:
\describe{
\item{stats_tri}{Function to be applied to compute statistics for polygons either
single or multiple inputs as character. Supported statistics are: "mean",
"median", "sd", "min", "max", "sum" "var".}
\item{engine}{The preferred processing functions from either one of "zonal",
"extract" or "exactextract" as character.}
}
}
\examples{
library(sf)
library(mapme.biodiversity)
(aoi <- system.file("extdata", "sierra_de_neiba_478140_2.gpkg", package = "mapme.biodiversity") \%>\%
  read_sf() \%>\%
  init_portfolio(
    years = 2000:2020,
    outdir = system.file("res", package = "mapme.biodiversity"),
    tmpdir = system.file("tmp", package = "mapme.biodiversity"),
    add_resources = FALSE,
    cores = 1,
    verbose = FALSE
  ) \%>\%
  get_resources("srtmdem") \%>\%
  calc_indicators("tri", stats_tri = c("mean", "median", "sd", "var"), engine = "extract") \%>\%
  tidyr::unnest(tri))
}
\references{
Riley, S. J., DeGloria, S. D., & Elliot, R. (1999). Index that quantifies
topographic heterogeneity. intermountain Journal of sciences, 5(1-4), 23-27.
}
\keyword{indicator}
