% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_and_remove}
\alias{unzip_and_remove}
\title{Helper to unzip and remove zip/gzip files}
\usage{
unzip_and_remove(zip = NULL, dir = tempdir(), remove = TRUE)
}
\arguments{
\item{zip}{zip file to unzip}

\item{dir}{A directory where the extracted files are written to.}

\item{remove}{if TRUE, removes the zip else keeps it}
}
\description{
Use this function to unzip a zip/gzip file and remove the original
archive, if required.
}
\keyword{utils}
